Replace all chevron button implementation with common component
